+++ Tests/StreamChatTests/Controllers/ChannelController/ChannelController+SwiftUI_Tests.swift
@@ -97,6 +97,7 @@ final class ChannelController_SwiftUI_Tests: iOS13TestCase {
             deactivatedAt: nil,
             lastActiveAt: .unique,
             teams: [],
+            language: nil,
             extraData: [:]
         )